Update fn_markCustomObjs.sqf
Previous code would mark anything that inherits from 'buildings' class that is not 'HeliH' subtype, which would end up marking small objects. This checks the size of the bounding sphere of an object to see if it is large enough to deserve a mark. 8 seemed to work well as a min bounding sphere value for an object, but can be changed to include/exclude other objects. Also shortened the function some by adding shared code to in-file function and increased the marker alpha.